Create collection as seperate class
